Add "dbg" to better understand system.file()
extdataFile() returned "" and I could not find
out why. The problem was that system.file()
when being called without lib.loc, it looks for
the package and somehow returned the packages's
RStudio project folder instead of the folder
where it is installed to. There, there is no
direct "extdata" folder. It is within the
"inst" folder. Now, I set lib.loc explicitly
to .libPaths().